+++ src/main/java/paintingcanvas/animation/MovementAnimation.java
@@ -19,7 +19,7 @@ public class MovementAnimation extends Animation {
     @Override
     void updateAnimation(Drawable drawable, int frame, int duration) {
         if (frame == 0 || start == null) this.start = new Point(drawable.x, drawable.y);
-        var t = (float) easing.ease(frame / (double) duration);
+        var t = (double) easing.ease(frame / (double) duration);
         drawable.x = (int) (start.x + (end.x - start.x) * t);
         drawable.y = (int) (start.y + (end.y - start.y) * t);
     }
